fixing typos; adding comments; moving error function to new file
